Reenable Sponsor processing from BillStats
I have discovered that there are some bills in which there are no LDSPON
files, and only have 1 sponsor. That sponsor data is given to us by LBDC
in BILLSTATS.
Previously I thought that we always received that information twice
(LDPSON and BILLSTAT). LDBC will be contacted about only receiving the
data in LDSPON files.