Remove uses of elasticsearch delete by query.
Apis that use the elasticsearch delete-by-query plugin were not working
partially due to the fact that the delete-by-query plugin was not
installed on the elasticsearch instances.  To simplify server
maintenance, I have replaced the two uses of delete-by-query.  The
calendar search dao usage was not necessary and was replaced by an id
delete.  The committee search dao usage was replaced with a query
followed by a bulk delete on the results.