Fix es term queries in committee search dao.
By default elasticsearch indexes fields after passing them through the
analyzer, which makes all text lowercase.  This causes term filters to
return no results if the original term has uppercase characters and the
original value is passed in to the query.  This commit addresses this
issue by lowercasing the inputs to term queries in the committee search
dao.