Add frag. and publish info to change log deletes.
Adds logic to the change log functions to include the fragment id and
fragment published date time to the log if the function is logging a
delete operation.

This information was easily available for
update/insert functions because it was available directly from the
recently created row.  For deletes it is harder because the frag. id and
published date time are not passed into the delete query.

The id and pub. date time are inferred based on the recently added
sobi_fragment field process_start_date_time which is set when a sobi
fragment starts processing.  The function selects the fragment with the
maximum process start time to infer the fragment that was processing
when the delete occurred.  There is the potential for mixups if multiple
fragments are processing concurrently, but this is currently prevented
in the processing code.

Fixes #11822