Preserve spacing in html bill text parsing.
The html bill text parser in BillTextUtils would normalize spacing for
text sections under <U> tags because a different method would be used to
retrieve the tag text than normal text.  This commit fixes this by using
the same code to extract the <U> text, and using a new boolean flag in
the recursive method to determine if text is a child of a <U> tag.