Perist html-stripped bill text
Performance tests on the xml branch code have indicated that bill
retrieval is much slower than in the dev branch.  We have concluded that
this is due to the fact that the xml branch stores bill text as html,
but strips the html when it is retrieved for a request.  This commit
will modify this stripping strategy so that the html is stripped during
lbdc data processing and stored in the database alongside the html text,
removing the need to perform stripping during a request.

- Add fullTextHtml field to BillAmendment
- Add full_text_html field to bill_amendment table with mapping to
  BillAmendment field in BillDao.
- In the XmlBillTextProcessor, strip html text and store in fullText
  field.  Html text is now stored in fullTextHtml field.
- Strip sponsor memo html before setting in the XmlSenMemoProcessor.
- Remove stripping methods from BillAmendment, replace usages with
  simple getters
- Update unit tests to properly verify new behavior.