Standardize ingest cache flushing
Ingest caching was extremely inconsistent. With some processors not even
calling the ingest cache, and with processors handling their own
implementations which would vary processor to processor.

By adding a checkIngestCache method in the SobiProcessor interface, this
forces all processors to implement something to check the cache. I then
wrote the method details to handle to corresponding caches.

We can be sure than nothing will be left in cache from not being flushed
anymore