Allow parsing of empty nysenate.gov bill votes.
A scenario was recently discovered where an nysenate.gov bill has vote
references that do not contain any vote data.  In this case, the bill
shows up in the bill dump with votes that have a value of "false".
These vote values could not be handled during parsing and caused an
exception to be raised.

This commit modifies the parser to handle this scenario.  Votes with
non-object values are ignored for parsing, causing them to be treated as
missing for the check.