Fix bug in Pipeline
The pipeline class has a bug which causes the pipeline future to be
completed when the first task is finished, as opposed to the final task
as it should be.  This occurs because the task futures are pushed to a
list instead of added, so when the last future in the list is selected,
it is the first task, not the last.

This is fixed by changed the push to an add.