Fix invalid Unicode character in xml Agenda parsng
When reporcessing sobis from the 2015 session, the agendas were failing
to parse because of the unicode character 0x1a. This character is not a
valid xml character and would cause a failure in processing. By simpling
replacing the character before it is passed to the xml helper this issue
no longer occurs