Restore Makeshift Session memeber method
In the xml environment, the newMakeshiftMember method was changed about
a year ago to just accept whatever was passed in and did not check the
input string with regex.
The way that sponsor information is processed
has changed in xml branch and it became the way we will process sobis
and xmls sponsor information. Adding the regex back solves issue 12066.
refs 12066