Fix bill retrieval from reprint info
In my most recent change to the billstat processor, I changed how the
basebill object is retrieved or created. I mistakenly used the reprint
information, instead of the bill information that the xml file is
published with. Additionally, the bill object has a bill id field
insteal of a base bill id for the reprint field. This is because
reprints have versions and not necessarily the last active version