Reduce mismatches loaded on spotcheck save.
When spotcheck reports are saved, a number of prior mismatches need to
be loaded in order to determine if any mismatches need to be closed and
the status of any observed mismatches.  The spotcheck report dao
currently loads all mismatches of a given type on save, which is
unnecessary, since only the mismatches for the observed content are
required.

This commit modifies the query that is used to retrieve the mismatches
for comparison.  The observed content ids are added to this query so
that only mismatches for the observed content are returned.