Modify updates api date range interpretation.
Modifies the updates apis so that they take open/closed ranges instead
of fully open ranges.  This makes it easier to poll for a continuous
range in smaller pieces.  Open/closed is used instead of the normal
convention of closed/open because in the most common use case, the user
is looking for updates that occurred after the last update that they
processed and it is simpler to provide that date to the api rather than
needing to increment it by a microsecond.