Fix bug in agenda view constructor.
In the constructor of AgendaCommAddendumView, the agenda id cannot be
set if the passed in AgendaInfoCommittee is null, even if there is a
valid AgendaVoteCommittee.  This is addressed by altering the addendumId
parameter to be a CommitteeAgendaAddendumId, which ensures that all of
the id fields can be set, even if info or votes are null.