Fix bug in saved spotcheck mismatch retrival.
There was a bug present in the query to retrieve spotcheck mismatches
for specific content keys for the purpose of determining resolved
mismatches for a report.  The query was attempting to match the string
value of key hstores with the output of the java dao function for
generating hstore strings.  There is no guarantee that these string
values will be equal, even if the java hstore generator creates values
that can be parsed into hstores.  To fix this, the query has been
modified so that passed in key parameters are converted to hstore and
then matched with stored keys as hstores.