Add vote roll mismatch to openleg bill spotcheck.
Add the vote roll mismatch to the openleg bill spotcheck.  This is a
check of simply which votes are present on a bill (id'ed by date, vote
type etc..) instead of an enumeration of the entire vote roll for every
vote of a bill.  This check is performed before the vote roll check, and
the vote roll check is skipped if a vote info mismatch is detected.