Fix Pipeline error handling.
The Pipeline class was not killing the pipeline task workers correctly.
The error handling method would use CompletableFuture#cancel to stop the
workers but this did not work as I had expected since CompletableFuture
does not have any control over the thread that is delivering the future.
I have successfully implemented the worker killing functionality by
setting a "finished" flag on the PipelineTask runnable objects.  This
flag is checked periodically during the PipelineTask excecution loop and
will break the loop if it is observed to be set.

Additionally, I have placed a 1 hour timeout on the pipeline execution
in the nysenate.gov bill report service.  This will prevent the report
from indefinitely blocking the data processor.

Fixes #12180