Fix xml approval memo removal.
The process for handling the "remove" action for approval xml files is
bugged.  It attempts to get a bill id parsed from the xml body, which
doesn't exist for "remove" files.  This is fixed by getting the bill id
from an existing approval memo with the same number, if one exists.

Also, the approval message removal unit test is improved.

Fixes #12177