Correct agenda vote type in xml whitelist
The XML whitelist was accepting the fragment type SENAGENV for agenda
vote data. However, XML agenda vote fragments have the type AGENDA_VOTE.

I have updated the whitelist and removed the SENAGENV since it is not
used anywhere. This is consistant with our use of AGENA over SENAGEN for
Agenda data.

refs #12148