Change test fragment handling for xml tests.
Modifies the way that xml processor tests process test fragments.  Prior
to this commit, each test class could test a single processor/fragment
type.  This commit sets up the base xml processor test to handle test
fragments in a way similar to the sobi process service, in which the
fragment type is determined by pattern matching in the text and the
correct sobi processor is selected from a map.  This enables us to write
tests that employ a combination of different fragment types/processors.