Transfer sponsors for new active amendment.
Modifies bill processor code to transfer sponsors when the active
amendment changes.  Sponsors are already copied when an amendment is
first created, but this was not enough in cases where new amendments
were not immediately published or previous amendments were restored as
active.  A unit test was added to test co/multi sponsor processing in
these cases.

Additionally some refactoring was done on bill processors:

- Consolidate xml and bill action parsing code because they parse
  actions in the same format.
- Reorganize xml billstat processor in an attempt to have a cleaner,
  easier to understand workflow.

Fixes #12151