Use console as default root log appender.
A previous commit added configuration for a file appender and set the
file appender as the appender for the root logger in the example log4j2
config.  This commit reverts it to use the console appender, since the
console appender is used in most cases i.e. all developers and prod.