Refactor ProcessConfig (again)
Rework the process config to fix some bugs, increase the number of xml
files used, and greatly simplify the logic.

The major bug addressed is that previously, the config was not
distinguishing between sobi and xml versions of fragments that are
produced by both data types, such as AGENDA and AGENDA_VOTE.  This was
because the white list filters used only filtered fragments based on
their fragment type.  This commit introduces filtering based on both
fragment type and the source type of the fragment, allowing the filters
to distinguish between sobi and xml versions of the shared fragment
types.  This also allows for simplification of the filter logic, which
can now be implemented as simple boolean expressions using fragment type
and source type.

Additionally, I have refined the filter selection logic and added more
filters to allow for a more precise filtering of files.  Instead of
selecting a filter by year, the applied filter is now selected based on
the sobi fragment/block's published date time.  This allows us to use a
good deal of 2017 xml data, and prevent usage of xml floor vote data
from before it was fixed in March of 2018.

Refs #12179