Fix agenda encoding issues
Fix an issue where apostrophes were incorrectly represented in agendas.
It seems that xml agendas are encoded in CP1252, so I added an exception
in the xml file dao to label agenda xmls as having this encoding.

Additionally, I have simplified the character substitution logic in the
agenda processor, and taken out some replacements that seemed to be
causing issues.

Fixes #12155, #12179