Revert text field custom index settings.
A previous commit added custom index settings to a number of
elasticsearch fields.  Many fields were changed to be the keyword type
to enable searching by default.  Unfortunately, this made it so that the
fields were not longer usable in wildcard searches by default.

This commit reverts the settings for these fields to their default
state.  A number of common text search fields are added to the utility
that handles sort strings and occurrences of these fields are now
converted to use the keyword subfield for ordering.

Additionally, the docs have been updated to instruct users to use the
keyword subfield for text searches.