Fix law document id used in search index.
For several years now we have been indexing laws based on their location
id, which is their location relative to the chapter of law.  Since all
chapters of law are placed in the same index, this causes a good deal of
overwriting.  This commit modifies the law search dao to use the full
law document id (which includes the law chapter) when indexing law
documents.

Refs #12194