Allow processing of externally scraped bills.
Modifies the dao and processor for scraped bills to enable processing of
bill html that was scraped on another instance.  This is convenient when
we have a limit of 1 scrape per minute.  This is accomplished by
breaking the scrape into two parts: the first where the html is saved to
a file in the incoming directory, and the second where the files in the
incoming directory are registered as scrape files in the db.