Fix bug in resolution header check.
In a previous commit I introduced a bug in the new resolution header
comparison added to the scraped bill check service.  When checking for a
header mismatch, the check service compared the reference text to itself
instead of the locally saved text, causing all resolution text
mismatches to be classified as header matches.  This commit fixes this
by changing the comparison to properly use both the reference and the
local texts.