Modify processing of BILLTEXTS for resolutions
BILLTEXT files can potentially reference up to two bills in the
billtext_html element attributes: one for senate and one for assembly.
Typically, if two bills are referenced, the text is applied to both.
When this occurs for resolutions, it is not valid because resolution
text contains the number and sponsor for the bill at the beginning, and
this number/sponsor can only be correct for one of the two bills.

This is fixed by applying resolution texts in a slightly different way:
only applying the text to the bill referenced in the print number parsed
from the file name.

Refs #12208