Fix sobi delete script bug from fkey removal
The recently updated delete_sobi_data.sql was not deleting agenda bill
and caledar data, because the foreign keys linking the tables for these
content types to the sobi fragment table were recently removed (but not
on my dev instance).  The script is updated to specifically target the
agenda, bill and calendar tables.