Creat missing budget bill text notification
A new notification that uses type WARNING has been created to alert us
of budget bill texts that are empty on the first version of the
amendment. It is set by default to run at 10am 1pm and 4pm everyday.
This requires a new app property that can be found in the example app
properties along with its default value
fixes #10916