Sameas uni resolutions text no longer sync
Xml Same as uni bills that are resolutions will no longer be allowed to
sync to the other houses bill text. This is due to the fact that
resolutions hae headers which are dependent on the house that it came
from. Overwriting an Assembly bill with a Senate header is wrong and
vice versa. A non resolution uni bill text will still sync as always