Batch load sobi fragments during processing.
Configures ManagedSobiProcessService to always load sobi fragments in
batches.  Previously, the number of fragments loaded was dependant on
the batch processing settings.  Since these settings are intended to
control the batching of the saves of sobi fragments and not how many
fragments are loaded at once, I am changing it to batch load fragments
even when batch processing is disabled.  Though fragments are always
loaded in batches, they are still processed and updated one by one if
batching is disabled.