Missing rule sponsor sql patch
Issue #12149 is not holding up the xml deployment to prod.
With that said, we still wanted to reflect the most accurate data even
if we don't have the latest billstats.
This sql file will patch our db to ensure that everything that is
supposed to be from the Rules Committee says so.
This patch needs to be run after reprocessing is completed