Account for lrs outages when scraping.
This commit alters the error handling for LRS scraping to better account
for LRS outages.  When an exception occurs that could be the result of
an LRS outage, the message will be logged at warning level.  Consecutive
scraping exceptions will be tracked, and if consecutive exceptions occur
for over 12 hours, it will start producing error notifications.