Fix session issues in commitee processor test
The committee processor integration test, XmlSenCommProcessorIT,
contained a bug in which it attempted to verify a processed committee
against the current version of that committee.  This was failing because
the current committee method retrieves the current committee for the
current session, and a version of the reference committee doesn't exist
for the curren session as of this commit.  This would have also failed
if the referenced committee was updated during the 2017 session.

To fix this, I have modified the test to verify using the committee
version from < 1 second after the published date time of the test xml
file.

I have also fixed a spelling error for some files related to the test
and modified some formatting in the test.