Fix "is amended" check in nysenate.gov spotcheck
Alter the "is amended" check in the nysenate.gov spotcheck to reflect
the logic on the website.  On the website, the flag is supposed to
indicate that the overall bill has one or more amendments (excluding the
"original" amendment).  Prior spotcheck logic checked whether the
active amendment differed from the selected amendment.

Fixes #12320