Fix agenda alert chair parsing regex.
Fix a regex bug in the agenda alert parser.  The chair regex was not
matching names with a hyphen because the hyphen was not properly
escaped in the regex.  The hyphen is now properly escaped in the regex.

Additionally, I modified the archive method in the agenda alert dao to
make it work when the alert archive directory doesn't exist yet.

Fixes #12345