Check all supplementals and active lists in calendar alert spotchecks
Updates the calendar alert spotcheck report to check each supplemental
and active list in the calendar when it runs.

Mostly this was accomplished by creating a SpotCheckObservation for each
supplemental and active list instead of only one per calendar. This
allows for precise identification of the content with the mismatch and
consistant behavior when getting new calendar alert data.

refs #11716