Convert 12/19 elastic api log sql script to flyway
Converts the 12/19/18 sql script that drops the request and response
tables to flyway.  This script was created in the elastic-api-log
branch, which had not been merged into the dev branch when the existing
scripts were converted to flyway.  The truncation step is removed and an
'if exists' clause is added for compatibility with instances that have
already upgraded to the latest dev code and run the script.
Also, removing the temporary rules sponsor fix script.