Use queue system for saving api request data.
Implement a queue system for saving api request data to elasticsearch to
prevent overuse of asynchronous threads.  Previously, each request would
employ a new thread to save the request data, which was causing
problems.  Now, request data is saved to an in-memory threadsafe queue,
which is cleared and persisted in bulk during a scheduled process.

Fixes #12255