Simplify committee data model in db.
Simplifies the committee data model in the database by removing several
redundant fields

The main area of simplification is the "reformed" column of the
committee_version table, which was a redundant value that matched the
"created" value of the next version's row.  The redundancy technically
made it more efficient to query for point in time data, but it added
complexity in the dao layer.

This commit refactors the dao to use just the "created" time for queries
and computing the "reformed" time using a subquery.  It also simplifies
some of the update logic and removes a number of unused queries and DAO
methods.

Fixes #12203