Attempt to remove all Sobi Frag references
In this huge list of changes, I have attempted to remove all instances
of sobi fragments and tried to leave sobis in the name only if the class
directtly dealt with sobis only. LegDataFragments will now be used
instead of sobi fragments.

The work isnt complete. The database and some sql queries still use sobi
fragments. All of that must be refactored.  refs #11312

NOTE: This commit changes some things to app properties, be sure to
update your app properties.