Fix array index issue in agenda report service
A recent modification to the BaseAgendaCheckReportService added logic
that filtered out certain references from the incoming reference list.
If all references were filtered out, this created an index exception
later on when the code attempts to access the first reference of the
list.  This was not a problem before because a ReferenceDataNotFoundEx
(which is checked in the spotcheck service) was thrown if no references
were found.

This commit fixes this issue by throwing a ReferenceDataNotFoundEx if
the filtered reference list is empty.