Fix LDSpon amendment transfer test.
There was an error a while ago in the amendment transfer test in the
LDSpon integration test suite related to a bill print number that was
more than 5 digits.  Instead of being fixed this test was commented out.

This commit fixes the amendment transfer test by changing the print no
used in test logic/data to a valid print no.