Move spotcheck util methods out of abstract class
Changes the BaseSpotCheckService class into the SpotCheckUtils class,
which presents the utility methods of the former as public.  Spotcheck
services no longer extend the base class.  They now just implement the
SpotCheckService interface and use the SpotCheckUtils class to provide
the methods previously inhereted from the base class.

This was done to allow more flexible use of these utility methods.  None
of the methods depend on the spotcheck service implementation so they
don't need to be coupled to it.