Fix bug allowing creation of negative session year
There was a bug in the SessionYear class that allowed for the
instantiation of a negative session year, even though the creation of a
negative session is explicitly forbidden.  When a year of 0 is provided,
it passes the non-negative validation and is then computed to a session
year of -1.  This has caused issues where a session year of -1 is saved,
and then causes an exception when it is loaded.

This is addressed in this commit by validating the computed session year
instead of the directly passed in session year.  Now, when 0 is passed
in, it will validate the computed value of -1 and throw an exception,
preventing negative sessions from being persisted with the processors.

Refs #12698