Modify pubhearing parser for new datetime format.
Recently, we have started receiving hearings with a slightly different
date/time format that breaks the current parser.  The dates in the new
format contain labels ("Date:" and "Time:") for the dates/times.

This commit adjusts the parser to be able to deal with these labels.
Also, a new test is added to test the parsing of the new date/time
formats.

Fixes #12704