Fix issues checking resolution bill text in scrape bill spotcheck
The scraped bill spotcheck report was not correctly checking/handling
cases of openleg having null text if the reference did have text.

First fix, if the openleg bill amendment text is null, set it to an
empty string to avoid NPE's when regex matching.

Additionally this removes the ensureTextFormatExists method which was
attempting to ensure we only compared html/plain text if necessary. However,
there were a few issues with it:

- When it correctly detected a missing text type, its thrown exception
was not handled at all. So this exception would repeat every report
until someone manually stopped it.

- We cant avoid checking text because its not present in openleg.
Sometimes openleg will be missing bill text and that is an important
mismatch that we want to report.