Alter mismatch save logic for very large reports
The spotcheck dao was producing errors when saving large reports (>
~32000 observations, a little less than max short value) due to a
limitation of the number of query parameters that can be sent for a
single query.

This commit refactors the report save process so that observations are
saved in batches to avoid exceeding the above mentioned limit.  Also,
some simplification of MismatchUtils.