Add Spotcheck for NYSenate.gov law data
Adds a spotcheck report that checks nysenate.gov law data.  The data is
parsed from a data dump in a similar format to other nysenate.gov dumps.
The law data is checked on several different levels to cover the
different fields in the data dump:
 - Law Tree based on parent id
 - Law document data
 - Law document relational data
 - List of all law chapters

The front end is updated to accommodate the new law spotcheck data.

Fixes #11998