Properly handle invalid offset parameter in search APIs
Previously, an offset parameter which was too large would throw an
IllegalArgumentException which was not caught by any of our API
exception handlers. This caused mass error notifications to be sent to
our admins. Additionally, the API user got no feedback as to what they
were doing wrong.

This commit changes the thrown exception to a newly created
InvalidSearchParamException. This exception is caught in our API
exception handlers so notifications are not sent and a useful response
is returned to the user.