Converts some caches to be limited by element size instead of heap size
Caches that are sized by bytes on heap have a performance impact in that
the size of the entire cache needs to be calculated routinely. This
performance impact could potentially be significant for some of our
caches which had object graphs of > 50,000 objects.

Analysing our logs, I noticed consistent spiking (avg 12s) in response
times when an ehcache object graph > 50,000 warning message was logged.

This commit converts some of our caches to be limited based on element
count instead of byte size to eliminate any performance impact.