Move hard coded cache sizes into configuration
The cache size for many of our smaller caches were hard coded instead of
being customizable in the app.properties file. This commits adds
configuration for these caches.

The Shiro cache is the only cache still hardcoded. It is coded quite
strangely and we cant inject a configuration value into it.