Improve query looking for missing budget bill pdfs
Adds a check of the publish status when querying for budget bills
missing pdfs.

This additional check was added because we received an empty bill text
for an amendment X in this file: 2020-04-02-03.53.30.472239_BILLTEXT_S07500X.XML.
Amendmet X does not exist for this bill, this was likely a mistake and
should not be showing up in our missing budget bill pdf report.