Remove makeshift member logic
The makeshift member logic created stub member records when unknown
shortnames were encountered. We would then review the new shortname
and manually correct the records as needed.

The problem with this was that in between creating a makshfit member
and correcting the data, our APIs would return incorrect data.

This change makes member data into lookup data that is manually added
through sql migrations. If an unknown shortname is encountered,
processing is stopped until the data can be added for that member and a
PROCESS_EXCEPTION notification is sent out to developers.

This change also makes development easier as every environment will have
full, correct, and consistent member data