Fixes issues with the notification digets implementation
There was a problem with the notification subscription digest
implementation. It used NotificationType categories such as 'ALL' as
real NotificationTypes instead of synonyms for the types they represent.
And since the digest rate limiting was based on the NotificationType, a
subscriber to 'ALL' types would essentially have every notification
delayed, even important ones.

To fix this, this commit removes NotificationType categories such as
'ALL', 'EXCEPTION', etc. Users must now explicitly subscribe to all the
notifications they want. It also adds a binary logarithm for determining
when to send notifications that are reoccurring often instead of a rate
limit.

NotificationDigests (collections of notifications that have occurred
over a period of time) have also been removed. No one currently uses
these, I can't think of any useful usecase, and they were adding a
bunch of complexity to the code.

refs #8983