Prevent tasks from being lost when thread pool fills up
This modifies the configuration of our ThreadPoolTaskExecutor by
removing the queue capacity and the max pool size (which does nothing
without the queue capacity).

The queue capacity was previously set to 100, this would cause new tasks
to be rejected if the queue was full, losing our data or potentially
skipping important processing tasks. To prevent tasks from being
rejected the queue capacity has been removed which makes it limitless.