Update CSS and implement section search
The section search feature is now working in its first implementation.
Meaning, if you search for a section number the text for that section
will now appear.
The section also needs to activate by clicking on it, which is not
working.
Additionally, different css and html has been added to help present this
page better / closer to production.