Fix a bug when searching for 5 digit bill numbers
When searching for a bill with a 5 digit print number our API was
incorrectly returning zero results.

An example API call: /api/3/bills/search?term=A10528

However, A10528 does exist, and an be found at other endpoints or at
/api/3/bills/search?term=10528

The cause was a bug in a regex which caused A10528 to be parsed as A1
with a session year of 0528.